+++ b0tweb/TorProxy.cs
@@ -22,7 +22,7 @@ namespace b0tweb
         /// </summary>
         public TorProxy()
         {
-            this._process = new Process();
+            this._process = null;
         }
 
         /// <summary>
@@ -31,6 +31,9 @@ namespace b0tweb
         /// </summary>
         public void Establish()
         {
+            if (Processes.IsRunning("tor")) return;
+
+            this._process = new Process();
             ProcessStartInfo info = new ProcessStartInfo();
             info.WindowStyle = ProcessWindowStyle.Hidden;
 
@@ -47,7 +50,7 @@ namespace b0tweb
         /// </summary>
         public void Disconnect()
         {
-            if (!this._process.HasExited)
+            if (this._process != null  && !this._process.HasExited)
             {
                 this._process.Kill();
             }
